Remove cache and keep only pending requests
