+++ .gitignore
@@ -1,6 +1,8 @@
 .DS_Store
 node_modules
 dist
+build
+*/**/build
 
 # Remove some common IDE working directories
 .idea
