+++ package.json
@@ -15,7 +15,6 @@
     "cors": "^2.8.5",
     "eslint": "^6.7.2",
     "express": "^4.17.1",
-    "memory-cache": "^0.2.0",
     "nodemon": "^2.0.7",
     "ts-node": "^9.1.1",
     "typescript": "^4.1.3"
