+++ src/api.ts
@@ -1,6 +1,6 @@
 import express from 'express';
 import snapshot from '@snapshot-labs/strategies';
-import { cacheMiddleware } from './cache';
+import { pendingRequestsHandler } from './helpers/pendingRequests';
 import scores, { blockNumByNetwork } from './scores';
 import { clone } from './utils';
 
@@ -21,7 +21,7 @@ router.get('/strategies', (req, res) => {
   res.json(strategies);
 });
 
-router.post('/scores', cacheMiddleware, async (req, res) => {
+router.post('/scores', pendingRequestsHandler, async (req, res) => {
   const { params } = req.body;
   const { space = '', network, snapshot = 'latest', strategies, addresses } = params;
